Refactor expander
