+++ frontend/lib/src/components/core/Block/Block.tsx
@@ -18,13 +18,12 @@ import React, { ReactElement } from "react"
 import { AutoSizer } from "react-virtualized"
 
 import { Block as BlockProto } from "@streamlit/lib/src/proto"
-import ExpandableProto = BlockProto.Expandable
 import { BlockNode, AppNode, ElementNode } from "@streamlit/lib/src/AppNode"
 import { getElementWidgetID } from "@streamlit/lib/src/util/utils"
-import withExpandable from "@streamlit/lib/src/hocs/withExpandable"
 import { Form } from "@streamlit/lib/src/components/widgets/Form"
 import Tabs, { TabProps } from "@streamlit/lib/src/components/elements/Tabs"
 import ChatMessage from "@streamlit/lib/src/components/elements/ChatMessage"
+import Expander from "@streamlit/lib/src/components/elements/Expander"
 
 import {
   BaseBlockProps,
@@ -39,8 +38,6 @@ import {
   StyledVerticalBlock,
   styledVerticalBlockWrapperStyles,
 } from "./styled-components"
-
-const ExpandableLayoutBlock = withExpandable(LayoutBlock)
 
 export interface BlockPropsWithoutWidth extends BaseBlockProps {
   node: BlockNode
@@ -73,20 +70,20 @@ const BlockNodeRenderer = (props: BlockPropsWithWidth): ReactElement => {
     props.scriptRunId
   )
 
-  let child: ReactElement
   const childProps = { ...props, ...{ node } }
+  const child: ReactElement = <LayoutBlock {...childProps} />
+
   if (node.deltaBlock.expandable) {
-    // Handle expandable blocks
-    const expandableProps = {
-      ...childProps,
-      empty: node.isEmpty,
-      isStale,
-      expandable: true,
-      ...(node.deltaBlock.expandable as ExpandableProto),
-    }
-    child = <ExpandableLayoutBlock {...expandableProps} />
-  } else {
-    child = <LayoutBlock {...childProps} />
+    return (
+      <Expander
+        empty={node.isEmpty}
+        isStale={isStale}
+        widgetsDisabled={props.widgetsDisabled}
+        element={node.deltaBlock.expandable as BlockProto.Expandable}
+      >
+        {child}
+      </Expander>
+    )
   }
 
   if (node.deltaBlock.type === "form") {
