+++ frontend/lib/src/components/elements/Expander/styled-components.ts
